+++ conference-backend/src/main/java/com/ubb/scalability/conference/repository/TalkRepository.java
@@ -3,5 +3,5 @@ package com.ubb.scalability.conference.repository;
 import com.ubb.scalability.conference.model.Talk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TalkRepository extends JpaRepository<Talk, Long> {
+public interface TalkRepository extends JpaRepository<Talk, Integer> {
 }
